+++ simulations/1D/couette_multithreaded.jl
@@ -25,7 +25,8 @@ function run(seed, T_wall, v_wall, L, ndens, nx, ppc, Δt, output_freq, n_timest
 
     # create our grid and BCs
     grid = Grid1DUniform(L, nx)
-    boundaries = MaxwellWalls1D(species_data, T_wall, T_wall, -v_wall, v_wall, 1.0, 1.0)
+    bc_list = (FullyDiffuseBC1D(species_data, 1, T_wall, [0.0, -v_wall, 0.0]),
+               FullyDiffuseBC1D(species_data, 1, T_wall, [0.0, v_wall, 0.0]))
 
     # split cell indices into chunks
     cell_indices = Vector(1:nx)
@@ -121,12 +122,12 @@ function run(seed, T_wall, v_wall, L, ndens, nx, ppc, Δt, output_freq, n_timest
             end
 
             if (t >= avg_start)
-                @timeit local_timer "convect + surface compute (t)" convect_particles!(rng_local, grid, boundaries,
+                @timeit local_timer "convect + surface compute (t)" convect_particles!(rng_local, grid, bc_list,
                                     particles_local[1], pia_local,
                                     1, species_data, surf_props_chunks[chunk_id], Δt)
             else
                 # we don't need to compute surface properties before we start averaging
-                @timeit local_timer "convect (t)" convect_particles!(rng_local, grid, boundaries,
+                @timeit local_timer "convect (t)" convect_particles!(rng_local, grid, bc_list,
                                     particles_local[1], pia_local,
                                     1, species_data, Δt)
             end