+++ src/abstract_types.jl
@@ -12,3 +12,10 @@ abstract type AbstractGrid end
 Abstract type that holds NetCDF-output related data for I/O
 """
 abstract type AbstractNCDataHolder end
+
+"""
+    AbstractBC
+
+Abstract type for boundary conditions
+"""
+abstract type AbstractBC end