+++ CHANGELOG.md
@@ -11,11 +11,13 @@ specified. `ParticleVector` now also has a `D` type parameter.
 * Removed unused `update_particle_indexer_new_lower_count!` function
 * NNLS merging functions now do not take `v_multipliers`, `n_rand_pairs`, and `centered_at_mean` as arguments, and do not use any fictitious particles
 * The function `check_speed_bounds` has been removed
-* NetCDF output now does not write species names to a variable, but rather to a global attribute as a single comma-separated string
+* NetCDF output now does not write species' names to a variable, but rather to a global attribute as a single comma-separated string
 * NetCDF PhysProps output now does not write total moments; an `NCDataHolderMoments` has been added for that purpose
 * PhysProps no longer stores any moment data and `compute_props_with_total_moments!` has been removed
 * Particle netCDF I/O simplified, skips output of position and cell data for 0-D particles
 * `load_electron_neutral_interactions` replaced by an `ElectronNeutralInteractions` constructor
+* `MaxwellWalls1D` removed
+* Boundary conditions now need to be passed to `convect_particles!` or `convect_particles_and_compute_cell!` as a `Tuple` of boundary conditions, one for each surface.
 
 ### New functionality
 * `squash_pia!` can be called less frequently, i.e. only before particle convection, as better tracking of last index
@@ -23,10 +25,14 @@ in particle arrays has been implemented. See documentaton on contiguous indexing
 * NetCDF output now allows to choose exact format and defalts to `NC_64BIT_OFFSET`, significantly speeding-up I/O
 * `compute_moment_scaling!`, `compute_moments!` functions to compute scaled total velocity moments
 * `NCDataHolderMoments` struct type added for I/O of computed moment data
+* More specialized boundary conditions added for 1D simulations: `FullyDiffuseBC1D`, `MaxwellWallBC1D`, `SpecularWallBC1D`
+* The `ParticleIndexerArray` type now directly stores number of cells and species it is tracking (`.n_cells`, `.n_species`)
 
 ### Misc
 * Added tests that check that for unexpected memory allocations
 * Various performance optimizations
+* Minor bug fixes
+* Documentation improvements
 
 ## v0.7.10
 * Documentation improvements